requirements.txt
coreapi==2.3.3
djangorestframework==3.12.4
Django==3.2.5
factory-boy==3.2.0
feedparser==6.0.2
Markdown==3.3.4
psycopg2-binary==2.8.6
pytest-django==4.1.0
selenium==3.141.0
uWSGI==2.0.19.1
